safari-600.7-branch/LayoutTests/fast/mediastream/RTCPeerConnection-createAnswer-expected.txt

Tests RTCPeerConnection createAnswer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.createAnswer(requestSucceeded1, requestFailed1); did not throw exception.
PASS createAnswer request succeeded.
PASS pc.setRemoteDescription(sessionDescription, setDescriptionSucceeded, setDescriptionFailed); did not throw exception.
PASS setRemoteDescription succeeded.
PASS pc.createAnswer(requestSucceeded2, requestFailed2, {}); threw exception Error: TypeMismatchError: DOM Exception 17.
PASS createAnswer request failed.
PASS errorReason.name is "Invalid createAnswer argument."
PASS pc.createAnswer(requestSucceeded3, requestFailed1, {requestIdentity:"ifconfigured"}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
